+++ CRM/Civirules/Trigger/Cron.php
@@ -31,5 +31,16 @@ abstract class CRM_Civirules_Trigger_Cron extends CRM_Civirules_Trigger {
     );
   }
 
+  /**
+   * Returns the name of the trigger data class.
+   *
+   * This function could be overridden in a child class.
+   *
+   * @return String
+   */
+  public function getTriggerDataClassName() {
+    return 'CRM_Civirules_TriggerData_Cron';
+  }
+
 
 }