+++ CRM/Civirules/Trigger/Post.php
@@ -12,6 +12,20 @@ class CRM_Civirules_Trigger_Post extends CRM_Civirules_Trigger {
 
   protected $op;
 
+  /**
+   * Returns the name of the trigger data class.
+   *
+   * This function could be overridden in a child class.
+   *
+   * @return String
+   */
+  public function getTriggerDataClassName() {
+    if ($this->op == 'edit') {
+      return 'CRM_Civirules_TriggerData_Edit';
+    }
+    return 'CRM_Civirules_TriggerData_Post';
+  }
+
   public function setTriggerId($triggerId) {
     parent::setTriggerId($triggerId);