+++ CRM/CivirulesConditions/Form/Membership/EndDateChanged.php
@@ -1,0 +1,53 @@
+<?php
+/**
+ * Class for CiviRules Condition Membership End Date Changed Form
+ *
+ * @author Jaap Jansma (CiviCooP) <jaap.jansma@civicoop.org>
+ * @license AGPL-3.0
+ */
+class CRM_CivirulesConditions_Form_Membership_EndDateChanged extends CRM_CivirulesConditions_Form_Form {
+
+  /**
+   * Overridden parent method to build form
+   *
+   * @access public
+   */
+  public function buildQuickForm() {
+    $this->add('hidden', 'rule_condition_id');
+
+    $this->add('checkbox', 'end_date_after_old_end_date', ts('New end date after the old end date'));
+
+    $this->addButtons(array(
+      array('type' => 'next', 'name' => ts('Save'), 'isDefault' => TRUE,),
+      array('type' => 'cancel', 'name' => ts('Cancel'))));
+  }
+
+  /**
+   * Overridden parent method to set default values
+   *
+   * @return array $defaultValues
+   * @access public
+   */
+  public function setDefaultValues() {
+    $defaultValues = parent::setDefaultValues();
+    $data = unserialize($this->ruleCondition->condition_params);
+    if (!empty($data['end_date_after_old_end_date'])) {
+      $defaultValues['end_date_after_old_end_date'] = $data['end_date_after_old_end_date'];
+    }
+    return $defaultValues;
+  }
+
+  /**
+   * Overridden parent method to process form data after submission
+   *
+   * @throws Exception when rule condition not found
+   * @access public
+   */
+  public function postProcess() {
+    $data['end_date_after_old_end_date'] = $this->_submitValues['end_date_after_old_end_date'];
+    $this->ruleCondition->condition_params = serialize($data);
+    $this->ruleCondition->save();
+    parent::postProcess();
+  }
+
+}