+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## Version 2.12 (not yet released)
 
 * Added operator 'Does not contain string' to Field Value condition.
+* Make sure we trigger rules after transaction has completed (see also issue #21)
 
 ## Version 2.11