+++ CHANGELOG.md
@@ -12,7 +12,8 @@
 * Added action: "Update Numeric Value"
 * Also check smartgroups for 'Contact in group' condition
 * Fixed issue #67 ("Event Type is" condition doesn't work)
-* Add condition 'Recurring payment processor' that checks which payment processor is linked to the recur 
+* Add condition "Recurring payment processor" that checks which payment processor is linked to the recur 
+* Add condition "Compare old Membership Status to new Membership Status".
 
 ## Version 2.11